Relocated class into utils package as recommended.
